pyproject.toml
[project]
name = "paper2saas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.21",
    "aiosqlite>=0.22.1",
    "arxiv>=2.3.1",
    "baidusearch>=1.0.3",
    "chonkie>=1.5.1",
    "ddgs>=9.10.0",
    "fastapi[standard]>=0.128.0",
    "firecrawl-py>=4.12.0",
    "mistralai>=1.10.0",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "psycopg2-binary>=2.9.10",
    "pycountry>=24.6.1",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
